+++ CHANGELOG.md
@@ -2,14 +2,20 @@
 
 ## Unreleased Changes
 
+- Replace `ChannelConfig` in `Channel::new` ([#511])
+
 ### FEATURES
 
 - Add support for streamlining releases ([#507])
 
 ### IMPROVEMENTS
 
+- Update to `tendermint-rs` v0.17.1 ([#517])
+
 
 [#507]: https://github.com/informalsystems/ibc-rs/issues/507
+[#511]: https://github.com/informalsystems/ibc-rs/pull/511
+[#517]: https://github.com/informalsystems/ibc-rs/issues/517
 
 ## v0.0.6
 *December 23, 2020*